requirements.txt
# 基础依赖
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# 文本处理
jieba>=0.42.1
nltk>=3.8.1
spacy>=3.6.0
regex>=2023.5.5

# 向量检索
faiss-cpu>=1.7.4; platform_system != "Windows"
faiss-gpu>=1.7.4; platform_system == "Windows"

# Web框架
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# 日志和监控
loguru>=0.7.0
prometheus-client>=0.17.0

# 测试和开发
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# 其他工具
python-dotenv>=1.0.0
PyYAML>=6.0.0
rich>=13.4.2
typer>=0.9.0